ml-requirements.txt
amply==0.1.4
docutils==0.17.1
joblib==1.0.1
libify==0.78
numpy==1.20.2
pandas==1.2.4
PuLP==2.4
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0

asgiref==3.3.1
autopep8==1.5.4
Django==3.1.3
django-cors-headers==3.6.0
djangorestframework==3.12.2
flake8==3.8.4
importlib-metadata==2.0.0
mccabe==0.6.1
psycopg2==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
python-dotenv==0.15.0
pytz==2020.4
sqlparse==0.4.1
toml==0.10.2
xlrd==2.0.1
XlsxWriter==1.3.7
xlwt==1.3.0
zipp==3.4.0


